updated domain level classes created UniversityService and its implementation created SQL queries to pull the data from the db created mappers

+++ src/main/java/com/university/university_application/db/entity/LectorEntity.java
@@ -11,6 +11,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "lectors")
@@ -28,5 +29,5 @@ public class LectorEntity {
     @Column(nullable = false)
     private Degree degree;
     @Column(precision = 8, scale = 2)
-    private double salary;
+    private BigDecimal salary;
 }
